Travis support
This should provide you the bits you need in order to get started
with the free tier of Travis CI.  I also changed phpunit.xml.dist
a bit,  most of the removed lines specify built-in default values.

Strict mode is on, ensuring tests have assertations and that they
run within 1 second. SyntaxCheck was removed in PHPUnit 3.6.